Adjust for size increases of some key struct members (See mathopd.h, rev 1.252.) Fix printf format warnings by using z and j modifiers.

+++ src/core.c
@@ -257,7 +257,7 @@ static void pool_adjust(struct pool *p, size_t s)
 	if (s > 0 && s < n)
 		n -= n % s;
 	if (debug)
-		log_d("pool_adjust: n=%d", n);
+		log_d("pool_adjust: n=%zu", n);
 	p->ceiling = p->floor + n;
 }
 
@@ -346,7 +346,7 @@ static int fill_connection(struct connection *cn)
 {
 	struct pool *p;
 	int poolleft, n, m;
-	long fileleft;
+	intmax_t fileleft;
 
 	if (cn->rfd == -1)
 		return 0;
@@ -359,7 +359,7 @@ static int fill_connection(struct connection *cn)
 	cn->left -= n;
 	m = read(cn->rfd, p->end, n);
 	if (debug)
-		log_d("fill_connection: %d %d %d %d", cn->rfd, p->end - p->floor, n, m);
+		log_d("fill_connection: %d %zd %d %d", cn->rfd, p->end - p->floor, n, m);
 	if (m != n) {
 		if (m == -1)
 			lerror("read");
@@ -422,7 +422,7 @@ static void write_connection(struct connection *cn)
 		}
 		m = write(cn->fd, p->start, n);
 		if (debug)
-			log_d("write_connection: %d %d %d %d", cn->fd, p->start - p->floor, n, m);
+			log_d("write_connection: %d %zd %d %d", cn->fd, p->start - p->floor, n, m);
 		if (m == -1) {
 			if (errno == EAGAIN)
 				return;
@@ -459,7 +459,7 @@ static int read_connection(struct connection *cn)
 	}
 	nr = read(cn->fd, cn->header_input.end, bytestoread);
 	if (debug)
-		log_d("read_connection: %d %d %d %d", cn->fd, cn->header_input.end - cn->header_input.floor, bytestoread, nr);
+		log_d("read_connection: %d %zd %zu %zd", cn->fd, cn->header_input.end - cn->header_input.floor, bytestoread, nr);
 	if (nr == -1) {
 		if (errno == EAGAIN)
 			return 0;
@@ -820,13 +820,13 @@ static void reap_children(void)
 			++stats.exited_children;
 			exitstatus = WEXITSTATUS(status);
 			if (exitstatus || debug)
-				log_d("child process %d exited with status %d", pid, exitstatus);
+				log_d("child process %d exited with status %d", (int) pid, exitstatus);
 		} else if (WIFSIGNALED(status)) {
 			++stats.exited_children;
 			termsig = WTERMSIG(status);
-			log_d("child process %d killed by signal %d", pid, termsig);
+			log_d("child process %d killed by signal %d", (int) pid, termsig);
 		} else
-			log_d("child process %d stopped!?", pid);
+			log_d("child process %d stopped!?", (int) pid);
 	}
 	if (pid < 0 && errno != ECHILD)
 		lerror("waitpid");
